backend/requirements/base.txt
Django>=4.2,<5.1
djangorestframework
channels[daphne]
django-environ
psycopg2-binary
celery
redis
django-cors-headers
